+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- auto-rotate session after 24 hours ([#247](https://github.com/PostHog/posthog-android/pull/247))
+
 ## 3.14.1 - 2025-04-23
 
 - fix: Send correct `$feature_flag_response` for the `$feature_flag_called` event when calling `isFeatureEnabled` ([#244](https://github.com/PostHog/posthog-android/pull/244))